Revert "profile set changes, remove banner beginnings"
This reverts commit e73fbab2db3a4e2ff7fd64e0fab69846512ca10a.
