+++ src/libs/ajax.js
@@ -133,8 +133,21 @@ const User = signal => ({
       return res.json()
     },
     set: keysAndValues => {
+      const blankProfile = {
+        firstName: 'N/A',
+        lastName: 'N/A',
+        title: 'N/A',
+        institute: 'N/A',
+        institutionalProgram: 'N/A',
+        programLocationCity: 'N/A',
+        programLocationState: 'N/A',
+        programLocationCountry: 'N/A',
+        pi: 'N/A',
+        nonProfitStatus: 'N/A'
+      }
       return fetchOrchestration(
-        'register/profile', _.mergeAll([authOpts(), jsonBody(keysAndValues), { signal, method: 'POST' }])
+        'register/profile',
+        _.mergeAll([authOpts(), jsonBody(_.merge(blankProfile, keysAndValues)), { signal, method: 'POST' }])
       )
     }
   },
